requirements.txt
Flask
Flask-WTF
openai>=1.0.0  # Use the latest version
python-dotenv